Repo hygiene wave: CI gates, lint policy, migration round-trip, route split, frontend tests (#718)
Cross-repo drift gate (SDKs + docs spec copy), pinned golangci-lint policy with 72 findings fixed or justified, coverage floor, migration round-trip in CI (six missing downs + 000007 fixed, orphaned migrations/ dir removed, org indexes shipped as 000175), /v1 route table extracted to routes_v1.go, ESLint 9 + 15 page test suites + UUID pickers, x/crypto CVE-2026-56854 bump, repo-local skills, and self-review fixes (interceptor header override, seed-tenant down, over-cap limit clamping, a11y association).

+++ cmd/seed/main.go
@@ -280,7 +280,7 @@ func main() {
 		// (Deferred -> Tax Payable), and — these are 'paid' — Code-3
 		// settlement (Cash -> AR). Same semantics the live posting path
 		// uses (internal/service/ledger.go).
-		total := int64(amount + 36000)
+		total := amount + 36000
 		if _, err := conn.Exec(`INSERT INTO ledger_transactions
 			(id, debit_account_id, credit_account_id, amount, ledger_id, code, reference_id, occurrence, description, created_at)
 			VALUES
